requirements.txt
# requirements.txt
opencv-contrib-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
mysql-connector-python==8.1.0
Pillow==10.0.0